+++ win/tclWinNotify.c
@@ -10,7 +10,7 @@
  * See the file "license.terms" for information on usage and redistribution of
  * this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * RCS: @(#) $Id: tclWinNotify.c,v 1.21 2005/11/04 00:06:50 dkf Exp $
+ * RCS: @(#) $Id: tclWinNotify.c,v 1.21.10.1 2010/05/21 12:18:17 nijtmans Exp $
  */
 
 #include "tclInt.h"
@@ -500,7 +500,7 @@ Tcl_WaitForEvent(
 
 	    PostQuitMessage((int) msg.wParam);
 	    status = -1;
-	} else if (result == -1) {
+	} else if (result == (DWORD)-1) {
 	    /*
 	     * We got an error from the system. I have no idea why this would
 	     * happen, so we'll just unwind.
